refactor: move sandbox kontinuation to mark it as a local function
